Use consistent data type
On iOS we need to run on both 32bit and 64 bit architectures, NSUInteger
differs depending on the architecture which causes issues when
interoping with the c# code.

+++ src/BugsnagUnity/Native/Cocoa/Breadcrumbs.cs
@@ -83,7 +83,7 @@ namespace BugsnagUnity
     }
 
     [MonoPInvokeCallback(typeof(NativeCode.BreadcrumbInformation))]
-    static void PopulateBreadcrumb(IntPtr instance, string name, string timestamp, string type, string[] keys, long keysSize, string[] values, long valuesSize)
+    static void PopulateBreadcrumb(IntPtr instance, string name, string timestamp, string type, string[] keys, int keysSize, string[] values, int valuesSize)
     {
       var handle = GCHandle.FromIntPtr(instance);
       if (handle.Target is List<Breadcrumb> breadcrumbs)
